Absorb more code into Maze::Client::Appium
