+++ bin/maze-runner
@@ -12,17 +12,17 @@ require_relative '../lib/maze/bitbar_utils'
 require_relative '../lib/maze/bitbar_devices'
 require_relative '../lib/maze/bugsnag_config'
 require_relative '../lib/maze/capabilities'
+require_relative '../lib/maze/client/bs_client_utils'
 require_relative '../lib/maze/client/appium/base_client'
 require_relative '../lib/maze/client/appium/bb_client'
 require_relative '../lib/maze/client/appium/bs_client'
 require_relative '../lib/maze/client/appium/local_client'
+require_relative '../lib/maze/client/appium/bs_devices'
 require_relative '../lib/maze/compare'
 require_relative '../lib/maze/docker'
 require_relative '../lib/maze/document_server'
 require_relative '../lib/maze/errors'
-require_relative '../lib/maze/farm/browser_stack/appium/devices'
 require_relative '../lib/maze/farm/browser_stack/capabilities'
-require_relative '../lib/maze/farm/browser_stack/utils'
 require_relative '../lib/maze/helper'
 require_relative '../lib/maze/logger'
 require_relative '../lib/maze/macos_utils'
